Reduce "Time out period for combos" from 50ms to 1ms.
(The setting was changed by going to QMK Settings / Combo in Vial.)

With the default value occasionally when I was typing words that contain "te" the keyboard would unexpectedly use the combo which generates "the ". This occurred infrequently but I don't want it to ever happen.

I've not been using the combos I defined.

A problem is that with the default timeout of 50ms that I couldn't always get the combo to work (rather than "because " I might get "bc", "cb", nothing or one of the letters - it only worked ~60% of the time). With a timeout of 100ms it still wasn't 100% reliable. (With 150ms it seemed to be reliable.) But with such a long timeout there's too much of a chance of the combo firing when I don't want it. (Can't win.)